newly envisioned JSON
This JSON more suits the direction of the project. I do have some issues like: I don't want the user to have to manually create a new JSON entry for a new cat. Meaning I need to think about how the JSON needs to be done, or if I should ditch json and have all the configuration done in-app (something like the debug menu)